Deprecate ProvidesEvents trait.
This trait was added before we had a naming convention for traits, and
has since been replaced by EventManagerAwareTrait. To help people
understand this, we should have the actual code in the correct trait,
and proxy the old one to the new one with a comment explaining what's
going on.
