+++ library/Zend/EventManager/ProvidesEvents.php
@@ -12,55 +12,11 @@ namespace Zend\EventManager;
 use Traversable;
 
 /**
- * A trait for objects that provide events
+ * NOTE: THIS TRAIT IS DEPRECATED! Please use EventManagerAwareTrait instead.
+ * This trait exists solely for backwards compatibility in the 2.x branch and
+ * willl likely be removed in 3.x.
  */
 trait ProvidesEvents
 {
-    /**
-     * @var EventManagerInterface
-     */
-    protected $events;
-
-    /**
-     * Set the event manager instance used by this context
-     *
-     * @param  EventManagerInterface $events
-     * @return mixed
-     */
-    public function setEventManager(EventManagerInterface $events)
-    {
-        $identifiers = array(__CLASS__, get_class($this));
-        if (isset($this->eventIdentifier)) {
-            if ((is_string($this->eventIdentifier))
-                || (is_array($this->eventIdentifier))
-                || ($this->eventIdentifier instanceof Traversable)
-            ) {
-                $identifiers = array_unique(array_merge($identifiers, (array) $this->eventIdentifier));
-            } elseif (is_object($this->eventIdentifier)) {
-                $identifiers[] = $this->eventIdentifier;
-            }
-            // silently ignore invalid eventIdentifier types
-        }
-        $events->setIdentifiers($identifiers);
-        $this->events = $events;
-        if (method_exists($this, 'attachDefaultListeners')) {
-            $this->attachDefaultListeners();
-        }
-        return $this;
-    }
-
-    /**
-     * Retrieve the event manager
-     *
-     * Lazy-loads an EventManager instance if none registered.
-     *
-     * @return EventManagerInterface
-     */
-    public function getEventManager()
-    {
-        if (!$this->events instanceof EventManagerInterface) {
-            $this->setEventManager(new EventManager());
-        }
-        return $this->events;
-    }
+    use EventManagerAwareTrait;
 }
